pyproject.toml
[project]
name = "s-ved-object-detection"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.2.2",
    "numpy>=1.18.5",
    "opencv-python>=4.1.2",
    "pandas>=1.1.4",
    "pillow>=7.1.2",
    "pyyaml>=5.3.1",
    "requests>=2.23.0",
    "scipy>=1.4.1",
    "seaborn>=0.11.0",
    "tensorboard>=2.4.1",
    "thop>=0.1.1.post2209072238",
    "torch>=1.7.0",
    "torchvision>=0.8.1",
    "tqdm>=4.41.0",
]
